move sync code from variable/mod.rs to variable/sync.rs
